좋아요 기능 exception 구체화

+++ src/main/kotlin/com/b05/newsfeedproject/domain/postslike/service/PostLikeServiceImpl.kt
@@ -1,5 +1,7 @@
 package com.b05.newsfeedproject.domain.postslike.service
 
+import com.b05.newsfeedproject.domain.exception.AccessDeniedException
+import com.b05.newsfeedproject.domain.exception.AlreadyExistException
 import com.b05.newsfeedproject.domain.exception.ModelNotFoundException
 import com.b05.newsfeedproject.domain.posts.repository.PostRepository
 import com.b05.newsfeedproject.domain.postslike.dto.PostLikeResponse
@@ -18,23 +20,32 @@ class PostLikeServiceImpl(
     override fun createPostLike(userId: Int, postId: Int) {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("post", postId)
         val user = userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("user", userId)
-        // post.userId 추가
-        return (0 != userId).let {
-            if (it)
-                postLikeRepository.findByUserIdAndPostId(userId, postId) ?: postLikeRepository.save(
-                    PostLike(
-                        post,
-                        user
-                    )
-                )
+
+        if (post.user.id == userId) {
+            throw AccessDeniedException("Your Post Like")
         }
+        if (postLikeRepository.findByUserIdAndPostId(userId, postId) == null) {
+            throw AlreadyExistException("Post Like")
+        }
+        postLikeRepository.save(
+            PostLike(
+                post,
+                user
+            )
+        )
     }
 
     override fun deletePostLike(userId: Int, postId: Int) {
+        if (postLikeRepository.findByUserIdAndPostId(userId, postId) == null) {
+            throw ModelNotFoundException("Post Like",postId)
+        }
         postLikeRepository.deleteByUserIdAndPostId(userId, postId)
     }
 
     override fun getPostLike(postId: Int): PostLikeResponse? {
-        return PostLikeResponse(postId, postLikeRepository.findByPostId(postId).size)
+        val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("post", postId)
+
+        return PostLikeResponse(postId, postLikeRepository.findByPostId(post.id!!).size)
     }
+
 }
